requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
python-multipart>=0.0.9
aiofiles>=23.2
pillow>=10.4
numpy>=1.26
requests>=2.31
google-generativeai>=0.8
openai>=1.0
ffmpeg-python>=0.2

# SEO + YouTube automation (phases 1-9)
cryptography>=42
google-auth>=2.29
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2
google-api-python-client>=2.125
feedparser>=6.0
APScheduler>=3.10
tenacity>=8.3
pydantic-settings>=2.2
httpx>=0.27
